Version 3.2:
Modifications to fix compatibility problem with old VELVCT code:
1) by default if called through VELVCT entry, vector text block is now
placed in bottom right corner of frame rather than under right corner
of viewport
2) if called through VELVCT, min and max vector sizes specified by FLO
and HI are now unconditional by default, meaning that they are not
dependent on the max/min magnitudes in the data.
VVINIT/VVECTR by default makes the max vector
size equal to the MIN(actual max vector in data, HI). But it is now possible
to choose the uncoditional mode using VVINIT/VVECTR as well by
specifying VHC and VLC as negative values.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1435 10fe2890-8a40-4696-8b1d-7ccbb19c398e