Added yet another type conversion interface - 	(But no one needs to use it except the new create/set/get interface)
Made all kinds of changes to the lowlevel resource setting functions in
preperation for the Fortran interface.

First cut an XyPlot - I need to clean some things up still.

Changed the Title object - the defaults for the "On" resources are now
dynamic.  If the user set's the "String" resources then the default value
for the "On" resources is True - If they don't set the "String" resources
the the default is False for the "On" resources.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1620 10fe2890-8a40-4696-8b1d-7ccbb19c398e