Checking in recent changes.
New constructs break and continue added. New operators and instructions
sequences for functions and procecures.

I merged the Functions.h and Procedures.h into a single header file.

Now use isatty to determine if input is coming from terminal. When
it is errors are recovered from immediately so the user can just
retype the bogus line. When the input is comming from a file or is
being loaded then error is not recovered from and parsing continues
but no execution results.

Implemented symbol table clean up functions that remove symbol
that were added in a bogus statement or block. This makes sure
that the runtime stack does not become fragemented.


ncl.l had a couple of bugs that resulted in the current line not
being saved for error reporting.

Made some mods to prompt is not written when no termnial is
input

-e


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1622 10fe2890-8a40-4696-8b1d-7ccbb19c398e