Added the SetValuesHook method to base.
Fixed some problems with the DataConversion w/regard to XyPlot.

XyPlot is "done" - I'm sure there are bugs, but it's time to start porting the
examples to use the new version, and to use the new data object.

Currently the only "DataItem" object that is ready to be used is the
coordArrTableLayerClass.  It is functional, but I haven't put in a setvalues
or getvalues method for it yet.  So you have to do everything from "Create".
But NhlAddData, NhlRemoveData both work with it, and you can even set the
NhlNcurveData resource of the XyPlot using a GenArray of DataItem id's
and it will work.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1623 10fe2890-8a40-4696-8b1d-7ccbb19c398e