Workstation.c: I modified the color map and marker table handling to use generic arrays; added a resource for the default foreground; changed the marker and fill implementation to cycle through the defined indexes using modular arithmetic if the index is out of range (in the positive direction).
hlu.c: I created a generic array copying function (_NhlValidatedGenArrayCopy)
that validates the contents of the array to be copied, by comparing it to a
target template GenArray.

Legend.c and LabelBar.c: Modified each of these objects to use the generic
Validated copy routine, and eliminated the local version. I also changed them
to take advantage of the modular capabilities of the fill and marker
routines.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1650 10fe2890-8a40-4696-8b1d-7ccbb19c398e