Added a ymkmf script that can be used to create a Makefile in any subdir so you don't have to make the entire Makefile tree to create a makefile in a part of the tree.
With no args, it parses the cwd and looks for the topdir named "ncarg".
It looks for the config dir in $topdir/config and will even try and
rebuild the ymake-filter if it is not executable.

It has an optional arg [topname] that it will use as the topdir if you
want to use ymake outside of the ncarg distribution.  If it is used this
way, ymake must be in your search path, and ymake will use $NCARG_ROOT
to find the config dir.

jeff

PS.  (I already installed the script in redcloud:/usr/local/ncarg/bin)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1666 10fe2890-8a40-4696-8b1d-7ccbb19c398e