TransObj modifications:
	The TransObj class now has a resource NhlNtrOutOfRangeF which
can be set to define the out of range value.

	All TransObj's now have mapping functions for all six mapping
methods (see TransObjP.h for field names of all six). Also now
each TransObj has all 4 line to functions defined.

	Changes to the mapping functions include: functions now replace
missing values and values that are out of the tranformations range to
the value of the resource NhlNtrOutOfRangeF. The function now set
the parameter *status to 1 if an out of range value is detected. The
status paremeter was previously called istrans and was set only when
a tranformation function was present, now since every TransObj must
have a function defined this parameter no longer had any meaning so
it could be reused for communicating when an ORV value is detected.

	MapTransObj forces the value of NhlNtrOutOfRangeF to 1e12 as
is required by EZMAP and anything used with EZMAP

Trans.c: All of the publicly callable private functions (_Nhl*) have
	been modified to call the internal methods and pass the status
	pointer in to the methods.

	The public functions have two new parameters one for the status
	which works like the internal calls and another float* for passing
	out the value of the ORV that replaces and missing or out of range
	values.

TickMark.c:
	Modified to use the new conventions for the the internal mapping
	TransObj methods.

	Also found a non-NULL terminated create. yikes!!

XyPlot.c:
	Modified to use the new conventions....

yMakefile:
	I put XyPlot.c and XyPlot.o back in as targets.

hlu.h:
	Interface changes to Trans.c functions


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1710 10fe2890-8a40-4696-8b1d-7ccbb19c398e