Added support for irregularly gridded data and fixed a number of miscellaneous bugs.
Contour: new routines for setting up an irregular tran obj, supporting code
throughout. Routine for drawing a bounding line into an areamap to
help prevent area fill "bleeding".

Overlay: Bug fixes in the ManageTickMarks routine

ScalarField: A number of new routines added to support data subsets,
strides, coordinate exchange, and irregular gridded data. The irregular
coordinate arrays are checked to ensure monotonicity and a one-to-one mapping
to the data array indexes.

Workspace: temporary debugging function call replaced with an internal
routine for dumping the contents of an areamap.

XyPlot: bug fix in the calculation of the center point of a LOG-scaled
"fake" irregular coordinate array.

hlu: Replaced the "NhlFree" call with "free" in the NhlTmpFile routine, since
the space is initially malloc'd by the system rather than the hlu library.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2129 10fe2890-8a40-4696-8b1d-7ccbb19c398e