I have installed a new internal parameter, named 'PIT' (for "Point Interpolation Threshold").  When mapping is turned on and 'PIT' is set non-zero, it causes the routines CPTRCL, CPTREG, and CPTRES (which trace contour lines, the edge of the grid, and the edge of the special-value areas, respectively) to look for adjacent mapped points which are larger than the value specified by 'PIT' and to interpolate points along the line segment being mapped so as to cause that line segment to be mapped more accurately.  This takes care of certain problems that arise (contour lines crossing other contour lines, for example).
While I was at it, I also improved the code in these routines that
looks for discontinuities in the mapping function.

I also fixed an error of long standing in CPTRCL that could, under
certain conditions, cause it to go into an infinite loop.  (I'm a
little surprised that this has never been reported by anyone.)

I have not yet run all the CONPACK tests, examples, etc., with the
new code.  I will do that this afternoon.  If you want to use the
new CONPACK code, check with me first to see if those tests have
turned up any problems.

I will be checking in a new example and new documentation for the
internal parameter 'PIT', but they're not done yet.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2131 10fe2890-8a40-4696-8b1d-7ccbb19c398e