The above files comprise an initial version of a package that I put together to do polygon manipulation.  Each of the files
  pp{di,in,un}{po,tr}.f

contains a routine that, given two polygons, performs the operation
indicated by characters 3 and 4 (difference, intersection, or union)
and then delivers either polygon boundaries or trapezoid boundaries
(depending on characters 5 and 6) defining the interior of the
resulting polygon to a user-defined processing routine.  These
routines are based on an article by Bala Vatti that was published
in the Communications of the ACM a couple of years ago and pointed
out to me by Fred.

The routine PPPLCL will clip a polyline against a rectangle and
deliver the resulting pieces of the polyline to a user-defined
processing routine.  This is done using the same algorithm that
Tom Wright put together - lo, these many years ago - for the old
package WINDOW.   I intend eventually to provide a routine that
will clip a polyline against an arbitrary polygon, but that routine
is not ready yet, so this is a temporary thing.

I actually wrote all these routines last year, but Fred only recently
decided to use a subset of them to do clipping in GKS.  At that point,
I subjected them to massive testing and now think they are ready for
use by whomever.

Shortly, I will check in a programmer document and an example for this
new package.  If interest warrants, I will put together "man" pages.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2218 10fe2890-8a40-4696-8b1d-7ccbb19c398e