This check in primarily represents additions to the MapTransObj and the ScalarField object, but a number of miscellaneous bug fixes and enhancements are also included.
Annotation.c: added code to ensure that an annotation unregisters itself
from the its overlay when it is destroyed.

Contour.c: Added converters for several enumerated types. Added
TickMark and Title management routines in order to support default
viewability of these annotations. Fixed a bug that caused the info label
not to appear on the first frame of a plot.

Converters.c: Fixed a bug in the GenArrayToFloatGenArray routine for
the NhlShort type. Added support for conversion of the NhlCharacter type.

MapPlot.c: Added a converter for the MapOutlineType enum, and beginnings of
support for a number of map style resources.

MapTransObj.c: Added converters for Projection and MapLimits types. Added
support for defining the map limits as Min/Max Lat/Lon with automatic setting
to the largest limits less than or equal to the requested values that
will not cause an Ezmap error. Also implemented
not totally foolproof error-checking for the Angle mode of defining
map limits.

Overlay.c: Added code to resize the default text heights of the titles
upon initialization based on the view size, when the user has not set
the heights explicitly. Added a converter for the AnnotationDisplayMode
enumeration.

ScalarField.c: Extensive rewrite of ScalarField to remove any dependency
on the datatype allowed for the field data. ScalarField now knows only about
its own type, float, and relies on the Converter mechanism to handle any other
types. Implemented the GetValues function for ScalarField.

Title.c: Fixed a bug that caused the wrong bounding box to be calculated
when the title is not actually displayed.

hlu.h: added new GenArray types to allow multi-dimensional genarray
conversion.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2235 10fe2890-8a40-4696-8b1d-7ccbb19c398e