Contour: Added support for formatting and scaling of all label types; rearranged some code for readability; added support for smoothing contour lines (not yet tested) and the new Conpack discontinuity testing parameter (also untested).
CoordArrTable: minor change to enable compile.

Workspace: Removed some confusing defines that looked like hlu data types
but were not. Replaced with a new enumerated type.

Format: implements a method for formatting floating point values using
a somewhat C-like format string. All Conpack floating point formatting
options are supported (including setting the left-most significant digit
based on a reference value such as the maximum value in the data set). A
number of other options are supported as well, such as choosing the
decimal point character, support for field widths, specifying the
decimal point position within the field, and choice of an arbitrary
character for filling the empty portion of a field. Note that some of
these options will require constant spacing to be turned on in order to
work properly. Format.c contains three global private functions --
	_NhlScanFstring parses the format string and returns an
NhlFormatRec with the structured formatting info;
	_NhlFormatFloat, given a format record and a floating point value,
as well as dynamic replacement values for some of the fields, returns
a formatted string.
	_NhlGetScaleInfo probably does not belong in this module,
but is somewhat related. Given a float point value it returns info that
is useful for scaling the number.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2301 10fe2890-8a40-4696-8b1d-7ccbb19c398e