Added full "ymake" support.
NclApi.c:	Just changed the management of the Error Object.  This file
		isn't even really used yet so it really doesn't matter.

NclFile.c:	Added a prototype for MyFileWriteVar to the beginning of
		this file.

yywrap.c:	Removed the reference to MAKEAPI and changed this file so
		it is only part of the EXE_OBJS.

yMakefile:	Very little remains the same.

To add addintional Data Objects to this management:

1.)	Create a script that is responsible for creating the .c file.
	It must be called ${TYPE}_data_c.sh.

2.)	Add a section to op_funcs.sh if necessary.

3.)	Create a file NclMultiDVal${TYPE}Data.c.specific with all the
	specific info in it.

4.)	Add a line to the end of the yMakefile using the MultiDValDataRule
	rule.


jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2327 10fe2890-8a40-4696-8b1d-7ccbb19c398e