Added resources to control the color of areas above the maximum specified level and below the minimum specified level. These resources apply only to the MANUAL and EXPLICIT modes of setting the contour levels.
Also modified the level-setting behavior in general. For AUTOMATIC and
EQUALSPACING modes the min and max data values are included as the
first and last contour levels. This simplifies the relationship between
contour lines and contour fill levels: there is now always one less fill
level than contour line level. The abovemax and belowmin resources are
independent of the level arrays. In other words they are not affected by
the various 'mono' flags.

Also changed the defaults for the resources that link together line, high
and low label attributes. Now the default is that each label type is
controlled individually.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2531 10fe2890-8a40-4696-8b1d-7ccbb19c398e