THIS IS FOR 4.0.
Only two of these routines have changed substantively: DPBLDA and DPDRAW.
In response to some problems that Fred came across, I have changed the way
in which I compute the length of a label.  Because of this, I have had to
change the default value of the internal parameter 'LS2' to 0.

I am checking in the other routines only to put a "$Id" on the front end
of them.  It turns out that they were created using "fsplit" on "cyclone",
which not only gives you files with upper-case names, but does not include
comments that lie between the end of one routine and the beginning of the
next, thus removing all my carefully-inserted "$Id"s.

Fred, you may find that you have to tweak your code a little, as the new
computation of the space required by a label includes a little white space.
I don't know how much it will affect you - probably not too much - but I
am going to have to tweak the DASHPACK example a little and you may have
to do the same in your code.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2540 10fe2890-8a40-4696-8b1d-7ccbb19c398e