I have removed all the stuff I previously added to do with quoting file paths.  The quoting broke other things.  Unfortunately, this means that directory names don't have any tokens in them, that happen to be defined for the cpp portion of ymake.  This means that using a directory name of sun4.4.1.1 will not work since the sun4 resolves to a "1" in cpp.  The way around this is simply to use _ underscores so that the sun4 is not a seperate token.  ie sun4_4_1_1.
In conjunction with this checkin, I have modified the ngdir script to return
appropriate directory names, and I have moved the dev directories to the
correct names.  I added sym-links to the dev directory from the old names
so any Makefiles, or .cshrc scripts that are currently referencing the
old names won't break.  These sym-links will be removed in a week or two.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2550 10fe2890-8a40-4696-8b1d-7ccbb19c398e