I am changing all function prototype ifdef's to use #if	NhlNeedProto so we don't have as many problems with mismatched prototypes.
Mary, it is very important that this library doesn't get installed without
everything that links to it also being recompiled.  Ncl and everything
needs to be recompiles and everything reinstalled at the same time.

This will cause acc -Xa on the Sun's to use the prototypes (it wasn't before).

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2746 10fe2890-8a40-4696-8b1d-7ccbb19c398e