DataSupport.c: 	Was not providing string name for logical type.
Execute.c:
	Was not properly passing coerced value when building subscripts. Also
was not freeing coerced value.
	Added execution sequence for new operator


NclFile.c:
	Several bug fixes. Found a bug that was causing NCL to read or write the
same data twice and on the second read/write and in correct error message was
generated. Several of the write functions were not checking the read/write
status of the opened file.

*specific:
	Added coercions to logical types for all type except strings,
characters,floats and doubles. Floats and doubles couldn't be done because
there is no way to map the missing value value for doubles into integers.

NclVar.c:
	Wasn't printing file variables and file variable subsections correctly.

	On named subscripting, the result array's named dimensions were not
in the correct order.

ncl.y:
	Changed precedence of exponential

OpsList.h:
	Added new operator to test if a variable is defined before building
	its subscript list. this allows ncl to provide a better error message
	when an identifier is undefined

Translate.c:
	Added new operator into translator


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2755 10fe2890-8a40-4696-8b1d-7ccbb19c398e