Rubber-banding code neglected to clip rubber band to be within device window. This change address the following bug report:
From haley  Tue May 10 14:46:14 1994
From: haley (Mary Haley)
Subject: idt/ictrans problem
To: clyne (John Clyne)
Date: Tue, 10 May 94 14:46:13 MDT
Cc: haley (Mary Haley)


Hi John,

I've got an idt problem but an error message is being echoed from ictrans,
so Jeff suggested I ask you about it.

I've got a metafile that has a plot in the upper left-hand quadrant only.
If you bring it up with idt and zoom on this plot so that it fills the
whole page, and then try to convert this plot to a PostScript using the
following ncarv_spool entry:

  myps : -d ps.mono : > gmeta.ps

you get an incomplete PostScript file (243 bytes).  But, if you zoom
on this full plot again (so you're in a sense just redrawing it), then
you get the correct PostScript output.  This seems to happen pretty
regularly on most machines (try it on redcloud using the idt in
/usr/local/ncarg/bin).  The main idt window reports the error message:

Display(0): Error - ctrans: FATAL - Could not process CGM element(class=0, id=1,
 name="Begin Metafile") [ Invalid window format


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2778 10fe2890-8a40-4696-8b1d-7ccbb19c398e