Contour: minor corrections
LabelBar and Legend: changed a number of resources documented as boolean
into the NhlBoolean type.

MapPlot: implemented scalar resources for array resources with the 'Mono'
flag; changed the names of many resources; added the ability to
modify the pattern and fill scale, in addition to color, to the
specified fill list; changed the name and behavior of the AreaGroupPriority
resource - it should be a bit simpler now. Also added an enumerated resource
that offers three options for managing the plot bounding box with respect
to the aspect ratio of the map.

MapTransObj: implemented new map limit modes: POINTS, NDC, and NPC.
NDC uses Normalized Device Coordinates to specify the limits; NPC are
"Normalized Projection Coordinates" - that is - the WINDOW limits mapped
into a 0.0 to 1.0 scale, allowing a uniform specification of a fraction
of the maximum projection space for all map projections. Also added a
resource for preserving the aspect ratio.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2881 10fe2890-8a40-4696-8b1d-7ccbb19c398e