This checkin fixes bugs reported in hlu00013, hlu00014, hlu00017, and hlu00025. As regards hlu00013, however, although the changes fix the problem in the example bug-report code, and hopefully many similar types of example, the root cause of the problem of contour bleeding and dropout, probably due to floating point error in the transform code, remains to be addressed. A number of other fixes are incorporated into this checkin as well.
Contour: the Areas parameter 'LC' is set to a smaller value, and
furthermore adjusted dynamically based on the viewport size. This fixes
the bug that shows up in hlu00013. Color management is also revised
slightly to adapt to changes to the Workstation colormap routines.

LabelBar: An ill-advised attempt to base the LabelBar orientation on the
aspect ratio of its viewport when it's not explicitly set was removed. This
fixes hlu00014. Some other dynamic resource setting code was improved.

Legend: Some dynamic resource setting code was improved.

MapPlot: An out-of-bounds array access was eliminated, curing hlu0025.
The name of the resource mpFillGroupCount was changed to mpAreaGroupCount.
This resource does not yet appear in any examples. Some color management
changes as well.

Overlay: fixed problem involving static variables that showed up when
attempting to make one Contour Overlay plot into an annotation of
another Contour Overlay plot.

Workstation: modified ColorMap management and access functions.
The wkColorMapLen now is the same as the highest allocated color index.
(Remember that the background color is *not* part of the colormap.) If you
do a GetValues of a colormap that contains 'holes' (unallocated color
indexes less than the highest allocated color index) all unallocated
color component entries contain the special value -1.0. If you set a
colormap with an array containing one or more negative color components,
those entries are considered to be unset and nothing is allocated.
The color index -1 (NhlTRANSPARENT) is considered to be an allocated color,
but of course, like the background color, it cannot be set using
NhlSetColor. There are now synonyms for the color index NhlTRANSPARENT and
the fill index NhlHOLLOWFILL: respectively NhlNULLCOLOR and NhlNULLFILL.
Although they can be used anywhere, they were implemented for the benefit
of MapPlot, because the names better describe the use of these indexes in
the mpSpecifiedFillColors and mpSpecifiedFillPatterns array resources.

Workspace: removed the debug printf that was appearing when the Map
areamap workspace increases in size.

bsearch, etc.: added RCS identifier tag.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@2970 10fe2890-8a40-4696-8b1d-7ccbb19c398e