I made some small changes to the Annotation interface. I have eliminated the NhlGetAnnotationId function since there is now a View resource that allows you to determine the Annotation id of any View object that is currently being used as an annotation. If the View is not an annotation the value of vpAnnotationId is 0 (NhlNULLOBJID). Otherwise it is the AnnotationId.
The Workstation object now is able to determine which View objects are
currently being used as annotations, and avoid drawing them directly
when you call Draw on the Workstation. It leaves the drawing up to
the Overlay Base Plot.

Also if you change the Workstation of an Overlay Base Plot, all Member
plots and annotations change to the identical Workstation.

Now if you destroy a View object that is currently used as an annotation,
it removes itself from the Overlay Plot before going away.

I changed the name of the Annotation resource anPlotId to anViewId, since
it's a more accurate name.

Rearranged where some functions are declared in order to reduce the
amount of unnecessary information being declared to the View, Annotation,
Reparent, and Workstation objects.

I added a check for setting out of bounds NDC values in the LLSetTrans
method. If out of bounds NDC values are encountered, the function now
issues a warning and constrains the plot to the viewspace.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3018 10fe2890-8a40-4696-8b1d-7ccbb19c398e