This is a much better fix for hlu00035.
Now, if the TextItem is a direct child of the Workstation object, it will
report a WARNING error if it has conflicting "vp"/"tx" positioning resources.
If it is being used within one of the other objects (LabelBar in the
case of hlu00035) it will just silently ignore the "vp" resources.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3110 10fe2890-8a40-4696-8b1d-7ccbb19c398e