THESE ARE BUG FIXES FOR 4.0.
Most of these changes are necessitated/enabled by some AREAS changes
that I'm about to check in.

In CPCLAM, I halved the values of a couple of constants that determine
how far apart the two tracings of the visible/invisible boundary that
go into the area map will be.  This leads to a cleaner-looking edge on
plots with such a boundary.

In CPTRCL, I fixed a bug (that crept in last fall), which was causing
this routine to sometimes generate unclosed contour lines.

In CPTREV and CPTRVE, I added code to return more accurate information
about the area identifiers on the contoured side of a visible/invisible
boundary line; this is important when the new way of reconciling area
identifiers is used.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3190 10fe2890-8a40-4696-8b1d-7ccbb19c398e