THESE ARE BUG FIXES FOR 4.0.
I have added a new internal parameter 'RC', which has the default value
0.  If it is given the value 1, contradictory area-identifier information
is reconciled in a new way: instead of using the most recent piece of
information seen, we use that value which we have seen most frequently
(ignoring zeroes).  If it is given the value 2, the same thing happens,
but zeroes are not ignored.  Using 'RC' = 1 seems to work *much* better
for contour plots than 'RC' = 0 and is *highly* recommended.

In the course of checking the new parameter out, I discovered and
fixed a subtle bug in ARPRAM; this bug may explain many of the cases
of "algorithm failure" and fill leakage that have occurred over the
years.  This is a *really* significant find.

I also added another capability to the debugging tools.  If you call
the routine "ARDBPA" with IGI zero or negative, it will show you *all*
the edges in the area map instead of just those in group IGI.  (You
can also set the internal parameter 'DB' negative to make ARPRAM call
ARDBPA in this fashion.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3191 10fe2890-8a40-4696-8b1d-7ccbb19c398e