Mary: this check-in should take care of HLU00032 and HLU00037.
Negative values of the ..ConstantSpacingF resources now cause a WARNING
error message, and are replaced with the default value 0.0.

Fixes to LabelBar and Legend: in AutoManage mode the text height is now
adjusted to compensate for changes to the constant spacing and aspect
ratio resources. A more geometrically accurate method of determining the
overlap of rotated labels was established. Although there is still some
change in the viewport size along the axis of orientation when the
label angle changes, it is much less pronounced. Changed the default
label offset values: lbLabelOffsetF now defaults to 0.1; lgLabelOffsetF
to 0.03.

ContourPlot.c: Several bug fixes involving line labels, and the way
line label resources are passed on to the Legend. Changed the default
line label placement mode to "computed".

MultiText.c: added a new resource: MtextMaxLenF: read-only resource
containing the maximum text length in the set of multi-text strings.
Removed some apparently unused code.

PlotManager.c: updated error messages to reflect new terminology.
changed the parameter names of the public functions to correspond to
their documented names.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3227 10fe2890-8a40-4696-8b1d-7ccbb19c398e