THIS IS FOR 4.0.
One more small extension:  Now you can use an IRC element which is -1 or -2.
The meaning is the same as if you had used a 1 or a 2, respectively, but, if
there are any negative area identifiers in the set of possibilities for a
given area, a -1 is used for the area.  (This is a response to a particular
test case that Ben Foster came up with.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3268 10fe2890-8a40-4696-8b1d-7ccbb19c398e