I'm checking these in so I can check them out on the sun.
These contain the infrasture to a callback system Jeff needs for
the GUI.

I also fixed bugs in the code for the ASSIGN_VAR_VAR_OP operator in
Execute.c that were causing the symbols to be left as undefined when
they infact referenced variables.

I modified the API function NclGetHLUObjId to return and NclExtValueRec
pointer.

The list_hlus function also had a problem when an array of type contained a
missing value.


-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3347 10fe2890-8a40-4696-8b1d-7ccbb19c398e