THIS IS FOR 4.0.
The object of this change is to remove some code that has been superseded
by recent bug fixes and is now more likely to cause trouble than to be of
help in avoiding trouble.

I think (hope) that this is the last change likely to result from my recent
review of AREAS/CONPACK area-fill problems.  (Of course, that sort of hubris
is guaranteed to lead to further problems ... :-)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3354 10fe2890-8a40-4696-8b1d-7ccbb19c398e