This check-in fixes bug #hlu00049.
ContourPlot.c: changed the default setting of 'cnSmoothingTensionF' to
be -2.5 rather than 2.5. This has the effect of causing smoothing to
take place prior to transformations of the data space. Otherwise
smoothing does not work properly when 'cnMaxPointDistanceF' is non-zero.
This will be documented. Actually a better solution might be to
add a boolean resource that controls when smoothing is applied explicitly.

I also fixed a bug that caused ContourPlot to use more levels than the
set value of 'cnMaxLevelCount' under certain circumstances.

LabelBar and Legend: I made a substantial number of parallel changes to
these two objects that fix some long-standing problems in the way the
geometry is set up. A user-visible change is that I renamed the
resources lb/lgMaxTitleExtentF to lb/lgTitleExtentF. These resources used
to set an upper bound on the amount of room the title was to occupy. If
less room was required LabelBar/Legend would try to use less space. Now
the resources simply specify as a fraction of the length of an axis minus
the margins, the amount of space the title is to occupy.

The other changes mostly apply to the extra calculations required when the
resources 'lb/lgAutoManage' is set False.

However, I also added code to allow the titles to be turned on automatically
any time the title string is set, and assuming 'lb/lgTitleOn' is not
explicitly set to False. This is consistent with the change Jeff recently
made to the Title class.


MapPlot: I added low-level error checking for all calls to the LLU library.

PlotManager: I added code that keeps the resources pmLabelBar/LegendWidthF
and pmLabelBar/LegendHeightF up to date. This is needed since both the
LabelBar and the Legend under certain circumstances
change their own viewports.

I will be updating the ref docs to reflect these changes.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3374 10fe2890-8a40-4696-8b1d-7ccbb19c398e