This checkin fixes bugs hlu00060 and hlu00062 (actually it does not do anything about the line label intervals, but that is not a bug, although it may be that the user should have more control of the Conpack parameters used to control the line label placement algorithms).
Specifically in hlu00060 the code now fills the areas the same whether or
not smoothing is turned on. There is a little corner that remains unfilled
in the upper right hand corner. I cannot seem to find a cure for this
however.

For hlu00062 I fixed a problem that prevented a number of the label font
color resources from being set to anything other than foreground or
background. There was no bug involving the cnGridBoundPerimColor.

Also I modified the areamap code along the lines of Dave Kennison's latest
recommendations. There now seems to be *much* less problem with unfilled
areas for contours transformed into map projections. I fixed some
problems involving the label masking resource.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3388 10fe2890-8a40-4696-8b1d-7ccbb19c398e