THIS IS FOR 4.0.
This is a new CONPACK example.  It produces four frames, the first of
which depicts a CONPACK-produced area map in great detail, and the
rest of which show situations in which problems might be expected.
The output is meant to be viewed while reading the descriptive text
that I have added to the programmer document ... which I will check
in next.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3393 10fe2890-8a40-4696-8b1d-7ccbb19c398e