This code checkin fixes bugs hlu00053 and ncl00037.
ContourPlot: Modified the default format to use 4 significant digits
instead of 6 (matching TickMark as well as Conpack defaults). The
left significant digit field of the MaxDataFormat resource, if set by
the user to 'dynamic', now set to the left significant digit value of
the maximum data value (as it should have been all along). The default
format also makes the exponential style use superscripts rather than
the 'E' format.

Converters: Slight modification to prevent a memory access 1 byte
beyond an array boundary that was detected by Purify.

Format: extended the syntax slightly to allow specific fields to be
explicitly specified as overrides of the 'set all numerics dynamic'
construct "*+". Modified the output generated for values equal to 0.0.
Zero values now contain the same number of significant digits as
other numbers generated using the same format. Simplified the logic and
fixed a bug involving zero used as a field fill character when
exponential notation is in effect. Modified the GetScaleInfo routine
to return a more accurate value for the number of significant digits
required to represent the subject value.

PlotManager: changed references to the Transform resource variable
overlay_on to its new name plot_manager_on.

TickMark: implemented a new method of formatting the numbers used as
TickMark labels based on the HLU floating point formatting specification.
More formatting options for TickMark labels are now available. The
default format spec causes the labels on each side all to use the same
number of digits following the decimal point. Certain options are
disabled when log coordinates are in effect. Also changed the code
so that a number of resources are kept up to date even when the
ticks for an axis are turned off. Previously not all resources were being
updated when an axis was off.

Transform: changed the variable name overlay_on to plot_manager_on,
making it more consistent with the name of the resource it applies to.

XyPlot: Fixed a bug in the code that recomputes the coordinate extents
based on the tick start and end. If the TickMark object is not
instantiated, the coordinate extents are no longer recomputed.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3454 10fe2890-8a40-4696-8b1d-7ccbb19c398e