This checkin fixes hlu00068.
MapPlot: added a private resource for dumping the area map for debugging
purposes.

TickMark: modified the DataLeft/Right and Bottom/Top resources to
default to 0.0 and 1.0. I also made changes so that if the two ends are
set to the same value the ticks for that side are turned off without
the indeterminate consequences evidenced in hlu00068.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3477 10fe2890-8a40-4696-8b1d-7ccbb19c398e