(The mnemonic "RWCHINFL" stands for "Read/Write CHaracters, INtegers, and FLoats".)
I am adding a new file of C I/O routines that I will be using to read
data in EZMAP and in PLOTCHAR.  They will also be used by a new version
of the routine used to convert the EZMAP outline dataset from ASCII to
binary form.

These routines have been tested on "parnassus", "cyclone", "alpine",
"shavano", "storm", "savage", and "chief" and seem to work just fine
everywhere.  I have tested new code for EZMAP on all of these machines,
as well, and will be checking that in tonight.  None of the new code
for PLOTCHAR is done yet, but it should be straightforward.

I'm going to have to rely on someone else to test some of these changes
in the context of an installation.  The routines in "rwchinfl.c" need
to be available at the time the EZMAP outline dataset is converted and
installed.  The conversion routine has been changed so that it now
creates four output files, rather than just one, and there are four
other files (eight in all) that need to be installed; I'm not sure
that I know how to update the yMakefile properly to reflect these
changes.

There will be several desirable effects of the updates that I'm doing:
one is to fix Ethan's problem of getting incorrect area ids when zooming
in too far.  Another is that EZMAP and PLOTCHAR will no longer use
FORTRAN logical units.  Another is that EZMAP and PLOTCHAR will be a
bit faster.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3511 10fe2890-8a40-4696-8b1d-7ccbb19c398e