These are the mods required to make PLOTCHAR read the PWRITX database using C I/O, rather than FORTRAN I/O.
I have tested the new code on "parnassus", "cyclone", "alpine", "shavano",
"storm", "savage", and "chief" and it seems to work just fine everywhere.
However, I have not actually tested the installation process on these
machines, except by doing "make all"s in this directory and in the directory
that creates the new version of the PWRITX database and observing that it
seems to do the right things.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3528 10fe2890-8a40-4696-8b1d-7ccbb19c398e