Here's the weird situation here:
With Solaris 2.4, whenever you link in the X11 library, it seems you also
have to link in the Xext library because of an undefined reference.  But,
the Xext library refers to routines in the X11 library, so you need to
include X11 twice, like this:

	-lX11 -lXext -lX11

This is a kludgy fix for 4.0.

Also, the Motif library resides in /usr/dt/lib, so I had to add this
library to the library search path.  The Motif library is only
available as a shared one, so I have to put "-Bdynamic" and "-Bstatic"
around -lXm.

All this makes it possible to compile "xproto" and other X applications
on a Solaris system.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3539 10fe2890-8a40-4696-8b1d-7ccbb19c398e